+++ SBVUnitTest/SBVUnitTestBuildTime.hs
@@ -2,4 +2,4 @@
 module SBVUnitTestBuildTime (buildTime) where
 
 buildTime :: String
-buildTime = "Sat Nov  7 17:54:01 PST 2015"
+buildTime = "Sat Nov  7 18:21:16 PST 2015"